pyproject.toml
[tool.poetry]
name = "pr-coach"
version = "0.1.0"
description = ""
authors = ["rmedvedo <rmedvedo@cisco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12.5,<=3.12.6"
crewai = "0.55.2"
langchain_ollama = "0.1.3"
crewai_tools = "0.12.0"
gitpython = "3.1.38"
PyGithub = "1.59.1"
langgraph = "0.2.34"
langchain_google_vertexai="1.0.10"
functions-framework = "^3.8.1"
google-cloud-logging = "^3.11.2"
anthropic="0.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
